basic version

+++ generate-dataset.js
@@ -2,15 +2,37 @@ const fs = require('fs-extra');
 
 const id = process.argv.pop();
 
-const project = require(`${__dirname}/projects/${id}/project.json`);
-const parse = require(`${__dirname}/projects/${id}/parser.js`);
+const run = id => {
+	return fs.exists(`${__dirname}/projects/${id}/parser.js`)
+		.then(exists => {
+			if (!exists) { 
+				console.log('skipping', id);
+				return;
+			}
+			const project = require(`${__dirname}/projects/${id}/project.json`);
+			const parse = require(`${__dirname}/projects/${id}/parser.js`);
 
-console.time('Parse');
+			console.log(`Parsing ${id}...`);
+			console.time(`Successfully parsed ${id}`);
+			
+			return parse()
+				.then(data => {
+					console.timeEnd(`Successfully parsed ${id}`);
+					Object.assign(project, data, { id });
+					let js = `window.projects.push(${JSON.stringify(project, null, '\t')});`;
+					js = js.replace(/\[[^\[\]]*\]/g, match => match.replace(/\s+/g, ' ')); // remove excessive whitespce
+					return fs.writeFile(`${__dirname}/build/projects/${id}.js`, js, 'utf8');
+				});
+		});
+
+};
+
 fs.ensureDir(`${__dirname}/build/projects`)
-	.then(() => parse())
-	.then(data => {
-		console.timeEnd('Parse');
-		Object.assign(project, data, { id });
-		let js = `window.projects.push(${JSON.stringify(project, null, '\t')});`;
-		return fs.writeFile(`${__dirname}/build/projects/${id}.js`, js, 'utf8');
+	.then(() => {
+		if (id && id.match(/^[A-Za-z0-9-]+$/)) {
+			run(id);
+		} else {
+			fs.readdir(`${__dirname}/projects`)
+				.then(projects => Promise.all(projects.map(id => run(id))));
+		}
 	});
